locking mutation - comment

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -96,7 +96,7 @@ const Mutation = {
         })
 
         if (!postExists)
-        throw new Error('unable to update post')
+            throw new Error('unable to update post')
 
         return prisma.mutation.updatePost({
             where: {
@@ -125,13 +125,15 @@ const Mutation = {
         }, info)
     },
 
-    async createComment(parent, args, {prisma}, info){
+    async createComment(parent, args, {prisma, request}, info){
+        const userId = getUserId(request);
+
         return prisma.mutation.createComment({
             data: {
                 text:args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -143,7 +145,18 @@ const Mutation = {
         },info)
     },
 
-    async deleteComment(parent, args, {prisma}, info){
+    async deleteComment(parent, args, {prisma, request}, info){
+        const userId = getUserId(request);
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists)
+            throw new Error('unable to delete comment')
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
@@ -151,7 +164,18 @@ const Mutation = {
         }, info)
     },
 
-    async updateComment(parent, args, {prisma}, info){
+    async updateComment(parent, args, {prisma, request}, info){
+        const userId = getUserId(request);
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists)
+            throw new Error('unable to update post')
+            
         return prisma.mutation.updateComment({
             where: {
                 id:args.id
